+++ vnag/__init__.py
@@ -21,4 +21,4 @@
 # SOFTWARE.
 
 
-__version__ = "0.4.0"
+__version__ = "0.6.0"